change TravelCertificateGraph.kt
